Tries to make something interactive, think its too finicky
