+++ 2021/24/input.js
@@ -1,12 +1,11 @@
-const path = require('path');
-const fs = require('fs');
+import path from 'path';
+import fs from 'fs';
+import { fileURLToPath } from 'url';
 
-const input = fs
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
+
+export const input = fs
 	.readFileSync(path.join(__dirname, 'input.txt'), 'utf8')
 	.toString()
 	.trim()
 	.split('\n');
-
-module.exports = {
-	input,
-};
