+++ 2021/24/part-one.html
@@ -1,0 +1,66 @@
+<!DOCTYPE html>
+<html lang="en">
+	<head>
+		<meta charset="UTF-8" />
+		<meta http-equiv="X-UA-Compatible" content="IE=edge" />
+		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
+		<title>Document</title>
+		<style>
+			input:invalid {
+				border: 2px dashed red;
+			}
+			input:invalid:focus {
+				background-color: pink;
+			}
+
+			input {
+				width: 200px;
+			}
+		</style>
+	</head>
+	<body>
+		<label>Input: <input id="num" placeholder="Enter 14 digit number" minlength="14" /></label>
+		<ul>
+			<li id="w">w: <span></span></li>
+			<li id="x">x: <span></span></li>
+			<li id="y">y: <span></span></li>
+			<li id="z">z: <span></span></li>
+		</ul>
+		<hr />
+		<pre id="winners-ele"></pre>
+		<script>
+			const input_ele = document.getElementById('num');
+			const winners_ele = document.getElementById('winners-ele');
+			(async () => {
+				const instructions_arr = await fetch('/input.txt')
+					.then((response) => response.text())
+					.then((raw_input) => raw_input.trim().split('\n'));
+
+				const { ALU } = await import('./alu.js');
+				const winners = new Set();
+
+				input_ele.addEventListener('input', () => {
+					const input_str = input_ele.value;
+					if (input_str.length === 14) {
+						const alu = new ALU(instructions_arr, input_str);
+						alu.run();
+						for (let register of ['w', 'x', 'y', 'z']) {
+							let output = document.querySelector(`#${register} > span`);
+							output.innerText = alu.registers[register];
+							if (register === 'z') {
+								if (alu.registers.z === 0) {
+									winners.add(input_str);
+									// Can sort by string since they are all the same length
+									winners_ele.value = [...winners].sort().join('\n');
+									output.innerText += ' ✅';
+								} else {
+									output.innerText += ' ‼️';
+								}
+							}
+						}
+					}
+				});
+			})();
+		</script>
+	</body>
+</html>
